+++ src/__generated__/Artwork_artwork.graphql.ts
@@ -5,11 +5,15 @@ import { AboutArtist_artwork$ref } from "./AboutArtist_artwork.graphql";
 import { ArtworkActions_artwork$ref } from "./ArtworkActions_artwork.graphql";
 import { ArtworkAvailability_artwork$ref } from "./ArtworkAvailability_artwork.graphql";
 import { ArtworkTombstone_artwork$ref } from "./ArtworkTombstone_artwork.graphql";
+import { ImageCarousel_images$ref } from "./ImageCarousel_images.graphql";
 import { OtherWorks_artwork$ref } from "./OtherWorks_artwork.graphql";
 import { SellerInfo_artwork$ref } from "./SellerInfo_artwork.graphql";
 declare const _Artwork_artwork$ref: unique symbol;
 export type Artwork_artwork$ref = typeof _Artwork_artwork$ref;
 export type Artwork_artwork = {
+    readonly images: ReadonlyArray<{
+        readonly " $fragmentRefs": ImageCarousel_images$ref;
+    } | null> | null;
     readonly " $fragmentRefs": ArtworkTombstone_artwork$ref & ArtworkActions_artwork$ref & ArtworkAvailability_artwork$ref & SellerInfo_artwork$ref & OtherWorks_artwork$ref & AboutArtist_artwork$ref;
     readonly " $refType": Artwork_artwork$ref;
 };
@@ -23,6 +27,22 @@ const node: ReaderFragment = {
   "metadata": null,
   "argumentDefinitions": [],
   "selections": [
+    {
+      "kind": "LinkedField",
+      "alias": null,
+      "name": "images",
+      "storageKey": null,
+      "args": null,
+      "concreteType": "Image",
+      "plural": true,
+      "selections": [
+        {
+          "kind": "FragmentSpread",
+          "name": "ImageCarousel_images",
+          "args": null
+        }
+      ]
+    },
     {
       "kind": "FragmentSpread",
       "name": "ArtworkTombstone_artwork",
@@ -55,5 +75,5 @@ const node: ReaderFragment = {
     }
   ]
 };
-(node as any).hash = '6b37bc5997cf3a6d13b8c3c1e96abb41';
+(node as any).hash = '230d7d14f29a59369fb609b37c5cc2ad';
 export default node;