+++ src/lib/Scenes/Artwork/__tests__/Artwork-tests.tsx
@@ -5,7 +5,7 @@ import { ArtworkActions } from "../Components/ArtworkActions"
 
 describe("Artwork", () => {
   it("renders a snapshot", () => {
-    const component = shallow(<Artwork artwork={{ " $fragmentRefs": null, " $refType": null }} />)
+    const component = shallow(<Artwork artwork={{ " $fragmentRefs": null, images: [], " $refType": null }} />)
     expect(component.find(ArtworkActions).length).toEqual(1)
   })
 })